Update 12/10/12
When the application launches, it first checks to see if there has
already been an entry recorded for today by fetching the most recent
entry in the core data stack. If there hasn't been, a timestamp is
recorded.

In terms of twitter integration, this should be working as-is, however
for some reason the application freezes when it should be asking the
user permissions to use twitter with the application. I'm looking into
this.

Other known bugs:
- Kal doens't call the markedDatesFrom: method, meaning we can't see
any marked dates. Still trying to figure out why this happens.